add setup script
add setup script modified from pypi example
