+++ hg-cache/cacheutils.py
